Custom Content and a Concept Level
The CustomStuff.upk consists mainly of textures, materials, and
speedtrees that I have made. The Johns_Ocean.upk file contains materials
that I have used as well. Both packages should be opened prior to
opening the test.udk level I have created. The test level is not yet
finished and is a work-in-progress concept design of what the game is
intended to look like.